TCOD for windows
A very smooth porting from curses to TCOD there. Port
was required since it became evident that WCurses
(Python curses for Windows) was pretty crappy.

Thanks to Antagonist on #rgrd for the assistance,
this file is basically a rip-off from part one of a
TCOD "how to make a roguelike" tutorial.